msys2 support (+cygwin +WSL)
More options for the configure script
MPI engine uses a type of the expected size (large enough to hold a pointer).
MPI assertaions only exists in MPI3.0.
Add the necessary libraries on Windows
Can't use INT on Windows, it will collide with system defined enum.
Enable support for rand_s
Support non-mmap reading of profiling file.
Correct line end on MINGW64 for the generated code.

+++ parsec/remote_dep.h
@@ -15,7 +15,7 @@
 
 BEGIN_C_DECLS
 
-typedef unsigned long remote_dep_datakey_t;
+typedef ptrdiff_t remote_dep_datakey_t;
 
 #define PARSEC_ACTION_DEPS_MASK                  0x00FFFFFF
 #define PARSEC_ACTION_RELEASE_LOCAL_DEPS         0x01000000
